fix the reinserting of cards when delete undo is called.
Details :
We used the old id of the deck to reinsert the card which changed after reinserting , so now we capture the newly inserted id and use that to reinsert the cards.

+++ app/src/main/java/com/adityap/flashy_createflashcards/dialogs/AddCardDialog.kt
@@ -57,7 +57,7 @@ class AddCardDialog : DialogFragment() {
 
                                 flashcardDatabaseHelper.addFlashcard(FlashcardModel(
                                         word = cardWord,
-                                        definition = cardDefinition), deckId)
+                                        definition = cardDefinition), deckId.toLong())
                                 listener?.onSave()
 
                                 dialog.dismiss()
